+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - [Expanded S3 API with versions and signed uri][14831]
 - [Support for reading JSON data from database connections.][14872]
 - [Add Redshift bulk loading][14860]
+- [JSON functions on Table and Column. XML and Hashset fixes.][14884]
 
 [14522]: https://github.com/enso-org/enso/pull/14522
 [14476]: https://github.com/enso-org/enso/pull/14476
@@ -64,6 +65,7 @@
 [14831]: https://github.com/enso-org/enso/pull/14831
 [14872]: https://github.com/enso-org/enso/pull/14872
 [14860]: https://github.com/enso-org/enso/pull/14860
+[14884]: https://github.com/enso-org/enso/pull/14884
 
 #### Enso Language & Runtime
 
@@ -75,6 +77,7 @@
 - [`polyglot java import` loads classes from HotSpot JVM][14798]
 - [Require opt-in to prefer local libraries][14885]
 - [One log file per execution][14893]
+- [Opt-in to get more detailed logs][14895]
 
 [14480]: https://github.com/enso-org/enso/pull/14480
 [14490]: https://github.com/enso-org/enso/pull/14490
@@ -84,6 +87,7 @@
 [14798]: https://github.com/enso-org/enso/pull/14798
 [14885]: https://github.com/enso-org/enso/pull/14885
 [14893]: https://github.com/enso-org/enso/pull/14893
+[14895]: https://github.com/enso-org/enso/pull/14895
 
 # Enso 2025.3